+++ example/lib/main.dart
@@ -28,7 +28,7 @@ class _MyHomePageState extends State<MyHomePage> {
   @override
   void initState() {
     super.initState();
-    availableMychannel = MyChannel();
+    availableMychannel = MyChannel()..initState();
   }
 
   @override