+++ lib/src/StateChannel.dart
@@ -10,8 +10,7 @@ import 'ChannelSignal.dart';
 ///    ...
 /// }
 /// ```
-
-abstract class StateChannel<S extends ChannelSignal> {
+abstract class StateChannel<S extends ChannelSignal>{
   ///A controller with the stream it controls.
   final StreamController<S> _streamController = StreamController<S>.broadcast();
 
@@ -23,7 +22,9 @@ abstract class StateChannel<S extends ChannelSignal> {
     _streamController.sink.add(signal);
   }
 
-  ///Closes the StateChannel.
+ initState(){}
+ 
+///Closes the StateChannel.
   void dispose() {
     _streamController.close();
   }