+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.1.0] - 2019.07.12
+
+* The initState method has been added to the StateChannel class.
+* SetState, initState, dispose methods have been added to the BaseState class.
+* 'signal' option has been added to wait, doneSucces, doneError methods of BaseState class.
+
 ## [0.0.3] - 2019.06.30
 
 * New description added.